Module with responsability to autenticate a user
