+++ backend/__tests__/integration/userRoutes.test.js
@@ -2,13 +2,12 @@
 
 const app = require('../../src/app')
 const request = require('supertest')
-const bcrypt = require('bcryptjs')
-
-const factory = require('../factories/userFactory');
 
 const { clearCollections, disconnect } = require('../utils/database');
+const { createUserWithDefinedPassword, getAutenticatedToken } = require('../utils/autenticatedUser')
 
 afterAll(async () => {
+    await clearCollections();
     await disconnect();
 })
 
@@ -51,21 +50,3 @@ describe('DELETE /users', () => {
         expect(body).toStrictEqual({ error: "Token não fornecido." })
     });
 });
-
-const getAutenticatedToken = async ({ email, password }) => {
-    const { body } = await request(app)
-        .post('/login')
-        .send({ email, password });
-
-    return body.token;
-}
-
-const createUserWithDefinedPassword = async ( password ) => {
-    const hash = await bcrypt.hash(password, 10);
-
-    const storedUser = await factory.create("User", {
-        password: hash
-    });
-
-    return { email: storedUser.email, password }
-}
